clean component
add avatar to form
remove utils from component
update logic for submit button
