+++ src/constants/userInfo.constants.js
@@ -1,0 +1,7 @@
+export const userInfoKeys = [
+  'userName',
+  'email',
+  'birthday',
+  'phone',
+  'telegram',
+];
